add the modSelection for the PrundDatabase
