+++ MetaMorpheus/TaskLayer/GlycoSearchTask/GlycoSearchParameters.cs
@@ -1,8 +1,9 @@
 using EngineLayer;
+using EngineLayer.GlycoSearch;
+using System.Collections.Generic;
+using System.IO.IsolatedStorage;
 using System.Linq;
 using UsefulProteomicsDatabases;
-using EngineLayer.GlycoSearch;
-using System.IO.IsolatedStorage;
 
 namespace TaskLayer
 {
@@ -33,6 +34,27 @@ namespace TaskLayer
             WriteContaminants = true;
             WriteSpectrumLibrary = false;
             DisposeOfFileWhenDone = true;
+
+            ModsToWriteSelection = new Dictionary<string, int>
+            {
+                //Key is modification type.
+
+                //Value is integer 0, 1, 2 and 3 interpreted as:
+                //   0:   Do not Write
+                //   1:   Write if in DB and Observed
+                //   2:   Write if in DB
+                //   3:   Write if Observed
+
+                {"N-linked glycosylation", 3},
+                {"O-linked glycosylation", 3},
+                {"Other glycosylation", 3},
+                {"Common Biological", 3},
+                {"Less Common", 3},
+                {"Metal", 3},
+                {"2+ nucleotide substitution", 3},
+                {"1 nucleotide substitution", 3},
+                {"UniProt", 2},
+            };
         }
         public string OGlycanDatabasefile { get; set; }
         public string NGlycanDatabasefile { get; set; }
@@ -57,6 +79,9 @@ namespace TaskLayer
         public bool WriteDecoys { get; set; }
         public bool WriteContaminants { get; set; }
         public bool WriteSpectrumLibrary { get; set; }
+        public bool WritePrunedDataBase { get; set; }
         public bool DisposeOfFileWhenDone { get; set; }
+
+        public Dictionary<string, int> ModsToWriteSelection { get; set; }
     }
 }
