+++ MetaMorpheus/TaskLayer/GlycoSearchTask/PostGlycoSearchAnalysisParameters.cs
@@ -4,6 +4,7 @@ using FlashLFQ;
 using Proteomics;
 using Proteomics.ProteolyticDigestion;
 using System.Collections.Generic;
+using Omics;
 using Omics.Modifications;
 
 namespace TaskLayer
@@ -24,5 +25,6 @@ namespace TaskLayer
         public List<string> CurrentRawFileList { get; set; }
         public FlashLfqResults FlashLfqResults { get; set; }
         public string IndividualResultsOutputFolder { get; set; }
+        public List<Protein> BioPolymerList { get; set; }
     }
 }
